requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
google-cloud-aiplatform==1.42.1
google-generativeai==0.3.2
Pillow==10.2.0
werkzeug==3.0.1
requests==2.31.0
Flask-Session==0.5.0
